another/backend/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
PyMySQL==1.0.3
flask-jwt-extended==4.4.4
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.0
scikit-learn==1.2.2
pandas==2.1.1
numpy==1.26.0
joblib==1.3.2
python-dotenv==1.0.1
bcrypt==4.0.1
gunicorn==20.1.0
cryptography
